refactor(alerts): move the alert error classifier into the shared platform
alert_error_classifier.py is a pure adapter over
posthog.errors.classify_query_error with no logs-specific content, and
its TRANSIENT_ERROR_CODES feeds the shared state machine's
is_transient_error. Moves it from products/logs/backend into
products/alerts/backend so a second adopter doesn't have to copy it.

Repoints logs' two importers and its test's patch targets at the new
location. No behavior change: same codes, same messages, same
transient set.
